Fixed issues with TS definitions
