+++ package.json
@@ -37,5 +37,6 @@
     "encoding": "^0.1.12",
     "fast-deep-equal": "^3.1.3",
     "isomorphic-unfetch": "^3.0.0"
-  }
+  },
+  "types": "./index.d.ts"
 }
