+++ react-native-bullet-train/index.d.ts
