+++ react-native-bullet-train/package.json
@@ -20,5 +20,6 @@
   "homepage": "https://bullet-train.io",
   "peerDependencies": {
     "react-native": ">=0.20.0"
-  }
+  },
+  "types": "./index.d.ts"
 }
